+++ model/src/main/java/dao/entities/ReturnNowPlayingSong.java
@@ -7,7 +7,7 @@ public class ReturnNowPlayingSong extends ReturnNowPlaying {
         this(other.getDiscordId(), other.getLastFMId(), other.getArtist(), other.getPlayNumber(), song);
     }
 
-    public ReturnNowPlayingSong(long discordId, String lastFMId, String artist, int playNumber, String song) {
+    public ReturnNowPlayingSong(long discordId, String lastFMId, String artist, long playNumber, String song) {
         super(discordId, lastFMId, artist, playNumber);
         this.song = song;
     }
